% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BA11.R
\docType{data}
\name{BA11}
\alias{BA11}
\title{Prices of 50,000 round cut diamonds.
Human postmortem brain data (Brodmann's area 11 in the prefronttal cortex)}
\format{
A data frame with 8325 rows and 293 variables. First row are time of death (TOD) of each subject and rest of rows are gene expression values. First column is gene symbol. For rest of columns, each column represents one individual.

\describe{
  \item{price}{price, in US dollars}
  \item{carat}{weight of the diamond, in carats}
  
}
}
\source{
\url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE71620}
}
\usage{
BA11
}
\description{
Prices of 50,000 round cut diamonds.
Human postmortem brain data (Brodmann's area 11 in the prefronttal cortex)
}
\references{
Chen, C.-Y., Logan, R. W., Ma, T., Lewis, D. A., Tseng, G. C., Sibille, E., and McClung, C. A. (2016).Effects of aging on circadian patterns of gene expression in the human prefrontal cortex.Proceedings of the National Academy of Sciences, 113(1):206–211.
}
\keyword{datasets}
