% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fisherInformation2.R
\name{fisherInformation2}
\alias{fisherInformation2}
\title{Fisher information matrix when two conditions exist}
\usage{
fisherInformation2(beta, tt1, yy1, tt2, yy2, period = 24)
}
\arguments{
\item{beta}{parameter vector of 8 with the following order: amp_1, phase_1, offset_1, theta_1, amp_2, phase_2, offset_2, theta_2}

\item{tt1}{time vector of condition 1}

\item{yy1}{expression vector of condition 1}

\item{tt2}{time vector of condition 2}

\item{yy2}{expression vector of condition 2}

\item{period}{Period of the since curve. Default is 24.}
}
\value{
The Fisher information matrix, this is a 8*8 matrix, with the same order as the input beta parameter.
}
\description{
Fisher information matrix when two conditions exist
}
\details{
Obtain the Fisher information matrix when two conditions exist
}
\examples{
set.seed(32608)
n <- 50
tt1 <- runif(n,0,24) 
Amp1 <- 2
Phase1 <- 6
Offset1 <- 3
yy1 <- Amp1 * sin(2*pi/24 * (tt1 + Phase1)) + Offset1 + rnorm(n,0,1)
tt2 <- runif(n,0,24) 
Amp2 <- 3
Phase2 <- 15
Offset2 <- 2
yy2 <- Amp2 * sin(2*pi/24 * (tt2 + Phase2)) + Offset2 + rnorm(n,0,1)
beta <- c(Amp1,Phase1,Offset1,1,Amp2,Phase2,Offset2,2)
fisherInformation2(beta, tt1, yy1, tt2, yy2)
}
\author{
Caleb
}
