% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FTest.R
\name{FTest}
\alias{FTest}
\title{F test for detecting circadian pattern.}
\usage{
FTest(tt, yy, period = 24)
}
\arguments{
\item{tt}{time vector}

\item{yy}{expression vector}

\item{period}{Period of the since curve. Default is 24.}
}
\value{
A list of amp, phase, offset, peak, SST, SSE, R2. 
Formula 1: \eqn{yy = amp \times sin(2\pi/period \times (phase + tt)) + offset}
Formula 2: \eqn{yy = A \times sin(2\pi/period \times tt) + B * cos(2*pi/period * tt) + offset}
\item{amp}{Amplitude based on formula 1}
\item{phase}{Phase based on formula 1, phase is restricted within (0, period)}
\item{offset}{Basal level(vertical shift) based on formula 1 or on formula 2}
\item{tss}{Total sum of square}
\item{rss}{Residual sum of square, SSE/n is the MLE of the variance sigma2}
\item{R2}{Pseudo R2 defined as (tss - rss)/tss}
}
\description{
F test sin function
}
\details{
Test the signficance of circadian curve fitting using F test
}
\examples{
set.seed(32608)
n <- 50
tt <- runif(n,0,24) 
Amp <- 2
Phase <- 6
Offset <- 3
yy <- Amp * sin(2*pi/24 * (tt + Phase)) + Offset + rnorm(n,0,1)
FTest(tt, yy)
}
\author{
Caleb
}
