% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LR_rhythmicity.R
\name{LR_rhythmicity}
\alias{LR_rhythmicity}
\title{Likelihood-based Tests for Detecting Circadian Pattern.}
\usage{
LR_rhythmicity(tt, yy, period = 24, method = "LR", type = "FN", B = 1000)
}
\arguments{
\item{tt}{Time vector}

\item{yy}{Expression vector}

\item{period}{Period of the since curve. Default is 24.}

\item{method}{Testing methods which can be "Wald", "LR", "F" or "Permutation". Default is "LR".}

\item{type}{Type of Test, finite sample "FN" or large sample "LS". Default is "FN". Only available if method = "LR" or "Wald".}

\item{B}{Number of permutations. Default is 1000. Only available if method = "Permutation".}
}
\value{
A list of amp, phase, offset, sigma02, sigmaA2, l0, l1, df, stat, and pvalue. 
Formula 1: \eqn{yy = amp * sin(2\pi/period * (phase + tt)) + offset}
Formula 2: \eqn{yy = A * sin(2\pi/period * tt) + B * cos(2*\pi/period * tt) + offset}
\item{amp}{Amplitude based on formula 1.}
\item{phase}{Phase based on formula 1, phase is restricted within (0, period).}
\item{offset}{Basal level (vertical shift) based on formula 1 or on formula 2.}
\item{sigma02}{Variance estimate under the null (intercept only).}
\item{sigmaA2}{Variance estimate under the alternative (since curve fitting).}
\item{l0}{Log likelihood under the null (intercept only).}
\item{l1}{Log likelihood under the alternative (since curve fitting).}
\item{stat}{Test statistic.}
\item{pvalue}{P-value from the test.}
\item{tss}{Total sum of square if available.}
\item{rss}{Residual sum of square, SSE/n is the MLE of the variance sigma2 if available.}
\item{R2}{Pseudo R2 defined as (tss - rss)/tss.}
\item{R2b}{Sequence of R squares for each permutaion in permutation test.}
}
\description{
Likelihood-based tests for circadian pattern detection.
}
\details{
Test the significance of circadian curve fitting using likelihood-based tests.
}
\examples{
set.seed(32608)
n <- 50
tt <- runif(n,0,24) 
Amp <- 2
Phase <- 6
Offset <- 3
yy <- Amp * sin(2*pi/24 * (tt + Phase)) + Offset + rnorm(n,0,1)
LR_rhythmicity(tt, yy, period=24, method="LR", type="FN")
}
\author{
Zhiguang Huo, Haocheng Ding
}
