% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LRTest.R
\name{LRTest}
\alias{LRTest}
\title{LRTest}
\usage{
LRTest(tt, yy, period = 24)
}
\arguments{
\item{tt}{time vector}

\item{yy}{expression vector}

\item{period}{Period of the since curve. Default is 24.}
}
\value{
A list of amp, phase, offset, peak, A, B, SST, SSE, R2. 
Formula 1: \eqn{yy = amp \times sin(2\pi/period \times (phase + tt)) + offset}
Formula 2: \eqn{yy = A \times sin(2\pi/period \times tt) + B * cos(2*pi/period * tt) + offset}
\item{amp}{Amplitude based on formula 1}
\item{phase}{phase based on formula 1}
\item{offset}{offset based on formula 1 or on formula 2}
\item{A}{A based on formula 2}
\item{B}{B based on formula 2}
\item{SST}{Total sum of square}
\item{SSE}{Error sum of square}
\item{R2}{Pseudo R2 defined as (SST - SSE)/SST}
}
\description{
Likelihood ratio test for circadian pattern detection
}
\details{
Test the signficance of circadian curve fitting using likelihood ratio test
}
\examples{
set.seed(32608)
n <- 50
tt <- runif(n,0,24) 
Amp <- 2
Phase <- 6
Offset <- 3
yy <- Amp * sin(2*pi/24 * (tt + Phase)) + Offset + rnorm(n,0,1)
LRTest(tt, yy)
}
\author{
Caleb
}
