% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LRTest_diff_amp.R
\name{LRTest_diff_amp}
\alias{LRTest_diff_amp}
\title{LRTest_diff_amp}
\usage{
LRTest_diff_amp(tt1, yy1, tt2, yy2, period = 24)
}
\arguments{
\item{tt1}{time vector of condition 1}

\item{yy1}{expression vector of condition 1}

\item{tt2}{time vector of condition 2}

\item{yy2}{expression vector of condition 2}

\item{period}{Period of the since curve. Default is 24.}
}
\value{
A list, see details below. 
Formula 1: \eqn{yy = amp \times sin(2\pi/period \times (phase + tt)) + offset}
Formula 2: \eqn{yy = A \times sin(2\pi/period \times tt) + B * cos(2*pi/period * tt) + offset}
\item{amp_1}{amplitute estimate of the 1st data}
\item{amp_2}{amplitute estimate of the 2nd data}
\item{amp_c}{amplitute estimate pooling all data together}
\item{l0}{log likelihood under the null (same variance between the two groups)}
\item{l1}{log likelihood under the alternative (different variance between the two groups)}
\item{df}{degree of freedom for the LR test}
\item{stat}{the LR statistics}
\item{pvalue}{the p-value from the LR test}
}
\description{
Likelihood ratio test for differential amplitute
}
\details{
Test differential amplitute of circadian curve fitting using likelihood ratio test
}
\examples{
set.seed(32608)
n <- 50
tt1 <- runif(n,0,24) 
Amp1 <- 2
Phase1 <- 6
Offset1 <- 3
yy1 <- Amp1 * sin(2*pi/24 * (tt1 + Phase1)) + Offset1 + rnorm(n,0,1)
tt2 <- runif(n,0,24) 
Amp2 <- 3
Phase2 <- 5
Offset2 <- 2
yy2 <- Amp2 * sin(2*pi/24 * (tt2 + Phase2)) + Offset2 + rnorm(n,0,1)
LRTest_diff_amp(tt1, yy1, tt2, yy2)
}
\author{
Caleb
}
