% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LRTest_diff_offset_FN_LS.R
\name{LRTest_diff_offset_FN_LS}
\alias{LRTest_diff_offset_FN_LS}
\title{LRTest_diff_offset_FN_LS}
\usage{
LRTest_diff_offset_FN_LS(tt1, yy1, tt2, yy2, period = 24, type = "LS")
}
\arguments{
\item{tt1}{time vector of condition 1}

\item{yy1}{expression vector of condition 1}

\item{tt2}{time vector of condition 2}

\item{yy2}{expression vector of condition 2}

\item{period}{Period of the since curve. Default is 24.}

\item{type}{Type of likelihood ratio test to use, "FN" or "LS". Default is large sample.}
}
\value{
A list, see details below. 
Formula 1: \eqn{yy = amp \times sin(2\pi/period \times (phase + tt)) + offset}
Formula 2: \eqn{yy = A \times sin(2\pi/period \times tt) + B * cos(2*pi/period * tt) + offset}
\item{offset_1}{offset estimate of the 1st data}
\item{offset_2}{offset estimate of the 2nd data}
\item{offset_c}{offset estimate pooling all data together}
\item{l0}{log likelihood under the null (same variance between the two groups)}
\item{l1}{log likelihood under the alternative (different variance between the two groups)}
\item{df}{degree of freedom for the LR test}
\item{stat}{the LR statistics}
\item{pvalue}{the p-value from the LR test}
}
\description{
Finite sample/Large sample Likelihood ratio test for differential offset.
}
\details{
Test differential offset of circadian curve fitting using likelihood ratio test
}
\examples{
set.seed(32608)
n <- 50
tt1 <- runif(n,0,24) 
Amp1 <- 2
Phase1 <- 6
Offset1 <- 3
yy1 <- Amp1 * sin(2*pi/24 * (tt1 + Phase1)) + Offset1 + rnorm(n,0,1)
tt2 <- runif(n,0,24) 
Amp2 <- 3
Phase2 <- 5
Offset2 <- 2
yy2 <- Amp2 * sin(2*pi/24 * (tt2 + Phase2)) + Offset2 + rnorm(n,0,1)
LRTest_diff_offset_FN_LS(tt1, yy1, tt2, yy2)
}
\author{
Caleb
}
