% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LRTest_diff_sigma2.R
\name{LRTest_diff_sigma2}
\alias{LRTest_diff_sigma2}
\title{LRTest_diff_sigma2}
\usage{
LRTest_diff_sigma2(tt1, yy1, tt2, yy2, period = 24)
}
\arguments{
\item{tt1}{time vector of condition 1}

\item{yy1}{expression vector of condition 1}

\item{tt2}{time vector of condition 2}

\item{yy2}{expression vector of condition 2}

\item{period}{Period of the since curve. Default is 24.}
}
\value{
A list of amp, phase, offset, peak, A, B, SST, SSE, R2. 
Formula 1: \eqn{yy = amp \times sin(2\pi/period \times (phase + tt)) + offset}
Formula 2: \eqn{yy = A \times sin(2\pi/period \times tt) + B * cos(2*pi/period * tt) + offset}
\item{amp}{Amplitude based on formula 1}
\item{phase}{phase based on formula 1}
\item{offset}{offset based on formula 1 or on formula 2}
\item{A}{A based on formula 2}
\item{B}{B based on formula 2}
\item{SST}{Total sum of square}
\item{SSE}{Error sum of square}
\item{R2}{Pseudo R2 defined as (SST - SSE)/SST}
}
\description{
Likelihood ratio test for sigma2
}
\details{
Test diff sigma2 of circadian curve fitting using likelihood ratio test
}
\examples{
set.seed(32608)
n <- 50
tt1 <- runif(n,0,24) 
Amp1 <- 2
Phase1 <- 6
Offset1 <- 3
yy1 <- Amp1 * sin(2*pi/24 * (tt1 + Phase1)) + Offset1 + rnorm(n,0,1)
tt2 <- runif(n,0,24) 
Amp2 <- 3
Phase2 <- 5
Offset2 <- 2
yy2 <- Amp2 * sin(2*pi/24 * (tt2 + Phase2)) + Offset2 + rnorm(n,0,1)
LRTest_diff_sigma2(tt1, yy1, tt2, yy2)
}
\author{
Caleb
}
