% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opt_commonSigma.R
\name{opt_commonSigma}
\alias{opt_commonSigma}
\title{fit the data based on joint sin curve assuming common sigmas}
\usage{
opt_commonSigma(tt1, yy1, tt2, yy2, period = 24)
}
\arguments{
\item{tt1}{time vector of condition 1}

\item{yy1}{expression vector of condition 1}

\item{tt2}{time vector of condition 2}

\item{yy2}{expression vector of condition 2}

\item{period}{Period of the since curve. Default is 24.}

\item{parStart}{initial value for optimzation purpose}
}
\value{
A list of amp, phase, offset, peak, A, B, SST, SSE, R2. 
Formula 1: \eqn{yy = amp \times sin(2\pi/period \times (phase + tt)) + offset}
Formula 2: \eqn{yy = A \times sin(2\pi/period \times tt) + B * cos(2*pi/period * tt) + offset}
\item{sigma2_1}{variance estimate of the 1st data}
\item{sigma2_2}{variance estimate of the 2nd data}
\item{sigma2_C}{variance estimate pooling all data together}
}
\description{
Optimization for common sigmas exist
}
\details{
Fit a joint sin curve for two conditions assuming common sigmas
}
\examples{
set.seed(32608)
n <- 50
tt1 <- runif(n,0,24) 
Amp1 <- 2
Phase1 <- 6
Offset1 <- 3
yy1 <- Amp1 * sin(2*pi/24 * (tt1 + Phase1)) + Offset1 + rnorm(n,0,1)
tt2 <- runif(n,0,24) 
Amp2 <- 3
Phase2 <- 5
Offset2 <- 2
yy2 <- Amp2 * sin(2*pi/24 * (tt2 + Phase2)) + Offset2 + rnorm(n,0,1)
opt_commonSigma(tt1, yy1, tt2, yy2)
}
\author{
Caleb
}
