% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PermTest_diff_phase.R
\name{PermTest_diff_phase}
\alias{PermTest_diff_phase}
\title{Permutation test for detecting differential phase}
\usage{
PermTest_diff_phase(tt1, yy1, tt2, yy2, B = 1000, period = 24)
}
\arguments{
\item{tt1}{Time vector for miltiple genes of condition 1}

\item{yy1}{Expression matrix for miltiple genes of condition 1}

\item{tt2}{Time vector for miltiple genes of condition 2}

\item{yy2}{Expression matrix for miltiple genes of condition 2}

\item{B}{Number of permutation for each gene.}

\item{period}{Period of the since curve. Default is 24.}
}
\value{
P-values, see details below. 
Formula 1: \eqn{yy = amp \times sin(2\pi/period \times (phase + tt)) + offset}
Formula 2: \eqn{yy = A \times sin(2\pi/period \times tt) + B * cos(2*pi/period * tt) + offset}
\item{pvalue}{P-value from the Permutation test}
}
\description{
Permutation test for differential phase
}
\details{
Test differential phase of circadian curve fitting using permutation test
}
\examples{
K=100
n=10
tt1 <- seq(2,2*n,2)
yy1 <- matrix(0,K,n)
tt2 <- seq(2,2*n,2)
yy2 <- matrix(0,K,n)
for(k in 1:K){
set.seed(k)
tt1 <- runif(n,0,24)
tt2 <- runif(n,0,24)
Amp1 <- runif(1,1,2)
Amp2 <- runif(1,3,4)
Phase1 <- runif(1,0,2)
Phase2 <- runif(1,5,7)
Offset1 <- runif(1,0,3)
Offset2 <- runif(1,5,8)
yy1[k,] <- Amp1 * sin(2*pi/24 * (tt1 + Phase1)) + Offset1 + rnorm(n,0,1)
yy2[k,] <- Amp2 * sin(2*pi/24 * (tt2 + Phase2)) + Offset2 + rnorm(n,0,3)
}
PermTest_diff_phase(tt1, yy1, tt2, yy2, B=1000)
}
\author{
Haocheng Ding
}
