% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PermTest_diff_R2.R
\name{PermTest_diff_R2}
\alias{PermTest_diff_R2}
\title{PermTest_diff_R2}
\usage{
PermTest_diff_R2(tt1, yy1, tt2, yy2, B = 1000, period = 24)
}
\arguments{
\item{tt1}{time matrix for miltiple genes of condition 1}

\item{yy1}{expression matrix for miltiple genes of condition 1}

\item{tt2}{time matrix for miltiple genes of condition 2}

\item{yy2}{expression matrix for miltiple genes of condition 2}

\item{B}{number of permutation for each gene.}

\item{period}{Period of the since curve. Default is 24.}
}
\value{
pvalues, see details below. 
Formula 1: \eqn{yy = amp \times sin(2\pi/period \times (phase + tt)) + offset}
Formula 2: \eqn{yy = A \times sin(2\pi/period \times tt) + B * cos(2*pi/period * tt) + offset}
\item{pvalue}{the p-value from the Permutation test}
}
\description{
Permutation test for differential R2
}
\details{
Test differential R2 of circadian curve fitting using Permutation test
}
\examples{
K=100
n=10
tt1 <- matrix(0,K,n)
yy1 <- matrix(0,K,n)
tt2 <- matrix(0,K,n)
yy2 <- matrix(0,K,n)
Amp1 <- 2
Phase1 <- 6
Offset1 <- 3
Amp2 <- 3
Phase2 <- 5
Offset2 <- 2
for(k in 1:K){
set.seed(k)
tt1[k,] <- runif(n,0,24) 
yy1[k,] <- Amp1 * sin(2*pi/24 * (tt1[k,] + Phase1)) + Offset1 + rnorm(n,0,1)
tt2[k,] <- runif(n,0,24) 
yy2[k,] <- Amp2 * sin(2*pi/24 * (tt2[k,] + Phase2)) + Offset2 + rnorm(n,0,5)
}
PermTest_diff_R2(tt1, yy1, tt2, yy2, B=1000)
}
\author{
Haocheng Ding
}
