% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pheno.R
\docType{data}
\name{pheno}
\alias{pheno}
\title{Subjects information for Brodmann’s Areas (BA) 11 Brain Aging Data}
\format{
A data frame with 146 rows and 3 columns. Each row represents one subject. First column is subject ID. Second row is age. Third row is time of death (TOD) in hours.
}
\source{
\url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE71620}
}
\usage{
pheno
}
\description{
Subjects information for human postmortem brain data (Brodmann's area 11 in the prefrontal cortex)
}
\references{
Chen, C.-Y., Logan, R. W., Ma, T., Lewis, D. A., Tseng, G. C., Sibille, E., and McClung, C. A. (2016).Effects of aging on circadian patterns of gene expression in the human prefrontal cortex.Proceedings of the National Academy of Sciences, 113(1):206–211.
}
\keyword{datasets}
