% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WaldTest.R
\name{WaldTest}
\alias{WaldTest}
\title{Wald test for detecting circadian pattern.}
\usage{
WaldTest(tt, yy, period = 24, type = "FN")
}
\arguments{
\item{tt}{Time vector}

\item{yy}{Expression vector}

\item{period}{Period of the since curve. Default is 24.}

\item{type}{Type of Test, finite sample "FN" or large sample "LS", default is "FN".}
}
\value{
A list of A, B, offset, df, stat, and pvalue
Formula 1: \eqn{yy = amp \times sin(2\pi/period \times (phase + tt)) + offset}
Formula 2: \eqn{yy = A \times sin(2\pi/period \times tt) + B * cos(2*pi/period * tt) + offset}
\item{A}{A based on formula 2}
\item{B}{B based on formula 2}
\item{amp}{Amplitude based on formula 1}
\item{phase}{Phase based on formula 1, phase is restricted within (0, period)}
\item{offset}{Basal level based on formula 1 or on formula 2}
\item{df}{Degree of freedom for the Wald test}
\item{stat}{Wald statistics}
\item{pvalue}{P-value from the Wald test}
\item{R2}{Pseudo R2 defined as (tss - rss)/tss}
}
\description{
Finite sample/ Large sample Wald test for circadian pattern detection
}
\details{
Test the significance of circadian curve fitting using finite sample Wald test
}
\examples{
set.seed(32608)
n <- 50
tt <- runif(n,0,24) 
Amp <- 2
Phase <- 6
Offset <- 3
yy <- Amp * sin(2*pi/24 * (tt + Phase)) + Offset + rnorm(n,0,1)
WaldTest(tt, yy)
}
\author{
Caleb
}
