% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WaldTest_diff_amp.R
\name{WaldTest_diff_amp}
\alias{WaldTest_diff_amp}
\title{Wald test for detecting differential amplitude}
\usage{
WaldTest_diff_amp(tt1, yy1, tt2, yy2, period = 24, type = "FN")
}
\arguments{
\item{tt1}{Time vector of condition 1}

\item{yy1}{Expression vector of condition 1}

\item{tt2}{Time vector of condition 2}

\item{yy2}{Expression vector of condition 2}

\item{period}{Period of the since curve. Default is 24.}

\item{type}{Type of Wald test to use, "FN" or "LS". Default is finite sample.}
}
\value{
A list, see details below. 
Formula 1: \eqn{yy = amp \times sin(2\pi/period \times (phase + tt)) + offset}
Formula 2: \eqn{yy = A \times sin(2\pi/period \times tt) + B * cos(2*pi/period * tt) + offset}
\item{amp_1}{Amplitude estimate of the 1st data}
\item{amp_2}{Amplitude estimate of the 2nd data}
\item{amp_c}{Amplitude estimate pooling all data together}
\item{df}{Degree of freedom for the Wald test}
\item{stat}{Wald statistics}
\item{pvalue}{P-value from the Wald test}
}
\description{
Finite sample/Large sample Wald test for differential amplitude.
}
\details{
Test differential amplitude of circadian curve fitting using Wald test
}
\examples{
set.seed(32608)
n <- 50
tt1 <- runif(n,0,24) 
Amp1 <- 2
Phase1 <- 6
Offset1 <- 3
yy1 <- Amp1 * sin(2*pi/24 * (tt1 + Phase1)) + Offset1 + rnorm(n,0,1)
tt2 <- runif(n,0,24) 
Amp2 <- 3
Phase2 <- 5
Offset2 <- 2
yy2 <- Amp2 * sin(2*pi/24 * (tt2 + Phase2)) + Offset2 + rnorm(n,0,1)
WaldTest_diff_amp(tt1, yy1, tt2, yy2)
}
\author{
Caleb
}
