% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WaldTest_diff_phase.R
\name{WaldTest_diff_phase}
\alias{WaldTest_diff_phase}
\title{Wald test for detecting differential phases.}
\usage{
WaldTest_diff_phase(tt1, yy1, tt2, yy2, period = 24, type = "FN")
}
\arguments{
\item{tt1}{time vector of condition 1}

\item{yy1}{expression vector of condition 1}

\item{tt2}{time vector of condition 2}

\item{yy2}{expression vector of condition 2}

\item{period}{Period of the since curve. Default is 24.}

\item{type}{Type of Wald test to use, "FN" or "LS". Default is finite sample.}
}
\value{
A list, see details below. 
Formula 1: \eqn{yy = amp \times sin(2\pi/period \times (phase + tt)) + offset}
Formula 2: \eqn{yy = A \times sin(2\pi/period \times tt) + B * cos(2*pi/period * tt) + offset}
\item{phase_1}{phase estimate of the 1st data, phase is restricted in (0, period)}
\item{phase_2}{phase estimate of the 2nd data, phase is restricted in (0, period)}
\item{phase_c}{phase estimate pooling all data together, phase is restricted in (0, period)}
\item{df}{degree of freedom for the Wald test}
\item{stat}{the Wald statistics}
\item{pvalue}{the p-value from the Wald test}
}
\description{
Finite sample/Large sample Wald test for differential phase.
}
\details{
Test differential phase of circadian curve fitting using Wald test
}
\examples{
set.seed(32608)
n <- 50
tt1 <- runif(n,0,24) 
Amp1 <- 2
Phase1 <- 6
Offset1 <- 3
yy1 <- Amp1 * sin(2*pi/24 * (tt1 + Phase1)) + Offset1 + rnorm(n,0,1)
tt2 <- runif(n,0,24) 
Amp2 <- 3
Phase2 <- 5
Offset2 <- 2
yy2 <- Amp2 * sin(2*pi/24 * (tt2 + Phase2)) + Offset2 + rnorm(n,0,1)
WaldTest_diff_phase(tt1, yy1, tt2, yy2)
}
\author{
Caleb
}
