% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BA11.R
\docType{data}
\name{BA11}
\alias{BA11}
\title{Brodmann’s Areas (BA) 11 Brain Aging Data}
\format{
A data frame with 8325 rows and 293 variables. First row is time of death (TOD) of each subject and rest of rows are gene expression values. First column is gene symbol. For rest of columns, each column represents one sample.
}
\source{
\url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE71620}
}
\usage{
BA11
}
\description{
Human postmortem brain data (Brodmann's area 11 in the prefrontal cortex)
}
\references{
Chen, C.-Y., Logan, R. W., Ma, T., Lewis, D. A., Tseng, G. C., Sibille, E., and McClung, C. A. (2016).Effects of aging on circadian patterns of gene expression in the human prefrontal cortex.Proceedings of the National Academy of Sciences, 113(1):206–211.
}
\keyword{datasets}
