% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitSinCurve.R
\name{fitSinCurve}
\alias{fitSinCurve}
\title{Fit Data Based on Sine Curve}
\usage{
fitSinCurve(
  tt,
  yy,
  period = 24,
  parStart = list(amp = 3, phase = 0, offset = 0)
)
}
\arguments{
\item{tt}{Time vector.}

\item{yy}{Expression vector.}

\item{period}{Period of the sine curve. Default is 24.}

\item{parStart}{Initial value for optimization purpose.}
}
\value{
A list of amp, phase, offset, peak, A, B, SST, SSE, R2. 
Formula 1: \eqn{yy = amp * sin(2\pi/period * (phase + tt)) + offset}
Formula 2: \eqn{yy = A * sin(2\pi/period \times tt) + B * cos(2*\pi/period * tt) + offset}
\item{amp}{Amplitude based on formula 1.}
\item{phase}{Phase based on formula 1, phase is restricted within (0, period).}
\item{offset}{Basal level (vertical shift) based on formula 1 or on formula 2.}
\item{A}{A based on formula 2.}
\item{B}{B based on formula 2.}
\item{tss}{Total sum of square.}
\item{rss}{Residual sum of square, SSE/n is the MLE of the variance sigma square.}
\item{R2}{Pseudo R2 defined as (tss - rss)/tss.}
}
\description{
Fit sin function
}
\details{
Fit a sine curve where tt is time, and yy is expression value.
}
\examples{
set.seed(32608)
n <- 50
tt <- runif(n,0,24) 
Amp <- 2
Phase <- 6
Offset <- 3
yy <- Amp * sin(2*pi/24 * (tt + Phase)) + Offset + rnorm(n,0,1)
fitSinCurve(tt, yy)
}
\author{
Caleb
}
