% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opt_commonOffset.R
\name{opt_commonOffset}
\alias{opt_commonOffset}
\title{fit the data based on joint sin curve assuming common offset}
\usage{
opt_commonOffset(tt1, yy1, tt2, yy2, period = 24, parStart = NULL)
}
\arguments{
\item{tt1}{time vector of condition 1}

\item{yy1}{expression vector of condition 1}

\item{tt2}{time vector of condition 2}

\item{yy2}{expression vector of condition 2}

\item{period}{Period of the since curve. Default is 24.}

\item{parStart}{initial value for optimzation purpose. This has the same order as the output vector.}
}
\value{
A vector of 7 with the following order: amp_1, phase_1, offset_c, theta_1, amp_2, phase_2,  theta_2
}
\description{
Optimization for common offset exist
}
\details{
Fit a joint sin curve for two conditions assuming common offset
}
\examples{
set.seed(32608)
n <- 50
tt1 <- runif(n,0,24) 
Amp1 <- 2
Phase1 <- 6
Offset1 <- 3
yy1 <- Amp1 * sin(2*pi/24 * (tt1 + Phase1)) + Offset1 + rnorm(n,0,1)
tt2 <- runif(n,0,24) 
Amp2 <- 3
Phase2 <- 15
Offset2 <- 2
yy2 <- Amp2 * sin(2*pi/24 * (tt2 + Phase2)) + Offset2 + rnorm(n,0,1)
opt_commonOffset(tt1, yy1, tt2, yy2)
}
\author{
Caleb
}
