% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WaldTest.R
\name{WaldTest}
\alias{WaldTest}
\title{WaldTest}
\usage{
WaldTest(tt, yy, period = 24)
}
\arguments{
\item{tt}{time vector}

\item{yy}{expression vector}

\item{period}{Period of the since curve. Default is 24.}
}
\value{
A list of A, B, offset, df, stat, and pvalue
Formula 1: \eqn{yy = amp \times sin(2\pi/period \times (phase + tt)) + offset}
Formula 2: \eqn{yy = A \times sin(2\pi/period \times tt) + B * cos(2*pi/period * tt) + offset}
\item{A}{A based on formula 2}
\item{B}{B based on formula 2}
\item{offset}{offset based on formula 1 or on formula 2}
\item{df}{degree of freedom for the Wald test}
\item{stat}{the Wald statistics}
\item{pvalue}{the p-value from the Wald test}
}
\description{
Wald test for circadian pattern detection
}
\details{
Test the signficance of circadian curve fitting using Wald test
}
\examples{
set.seed(32608)
n <- 50
tt <- runif(n,0,24) 
Amp <- 2
Phase <- 6
Offset <- 3
yy <- Amp * sin(2*pi/24 * (tt + Phase)) + Offset + rnorm(n,0,1)
WaldTest_finiteN(tt, yy)
}
\author{
Caleb
}
